Restructure using v1 programming model
