--- data_safe_haven/resources/gitea_mirror/functions/requirements.txt
@@ -1,6 +1,0 @@
-# Do not include azure-functions-worker in this file
-# The Python Worker is managed by the Azure Functions platform
-# Manually managing azure-functions-worker may cause unexpected issues
-
-azure-functions
-requests
